# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/maintenance_stages:
    get:
      tags:
        - MaintenanceStages
      summary: List MaintenanceStages
      description: Get a paginated list of MaintenanceStages with optional filtering and sorting
      operationId: listMaintenanceStages
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStageList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MaintenanceStages
      summary: Create MaintenanceStage
      description: Create a new MaintenanceStage
      operationId: createMaintenanceStage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceStageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceStage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    MaintenanceStage:
      type: object
      required:
        - id
        - name
        - sequence
        - fold
        - done
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
        sequence:
          type: integer
          format: int32
        fold:
          type: boolean
        done:
          type: boolean
        metadata:
          type: object
...
    CreateMaintenanceStageRequest:
      type: object
      required:
        - name
        - sequence
        - fold
        - done
        - metadata
      properties:
        company_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
        sequence:
          type: integer
          format: int32
        fold:
          type: boolean
        done:
          type: boolean
        metadata:
          type: object
...
    MaintenanceStageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceStage'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []